@mathiaswagner 2020-07-24
Should we query the requested tolerance? I.e. as I understand it even if you just request 1e-4 this would force double?