@weinbe2 2020-07-24
We discussed this and decided it was safest to always use double. The justification was it's not just the epsilon of the precision that matters, but also the conditioning of the matrix, so it's hard to just pick an epsilon and commit to it in a future+proof sense.